Don't clear the repo file if no file is provided on an update
If an update is saved by the user, and no file was specified, then the
file is reset to an empty file. The filename will be the empty string if
no file was specified. This changes the checks to work on the value of
the filename and not the value of the pointer.